Pin all upstream charms to current versions
Sadly this also requires to put back xenial/ everywhere, as a URL like
'cs:cinder-1' does not work, but 'cs:xenial/cinder-1' does.  I believe
this is because the version numbers are series-specific; for example the
current latest version for cinder on Xenial is 1, but for cinder on
Trusty is 253.