Use TLS to secure etcd connections
This commit uses an enhanced etcd charm (cs:~containers/etcd-4) that
requires TLS security for communication between the etcd master nodes,
and for connections from clients or etcd proxies.  The neutron-api and
neutron-calico charms, which both incorporate a local etcd proxy, have
been enhanced to establish the required TLS credentials for that etcd
proxy so that it can securely communicate with the etcd cluster.